… tensorflow/core/kernels/image/resize_op_benchmark_test.cc