This PR speedup `Resize{Bilinear,NearestNeighbor}` with eigen tensor generator class. The original implementation specialize 3-channeled image with SSE, but it's still less performant compared with this PR. Below is my benchmark with `tensorflow/core/kernels/image/resize_op_benchmark_test.cc`.

On ***Intel(R) Core(TM) i5-9600K CPU @ 3.70GHz***

### Nearest neighbor

| input_shape (N x IH x IW x C) | output_shape (OH x OW) | Old (M items/s) | New (M items/s) |
| ----- | ---- | ---- | ----- |
|  10x499x499x1       | 250x250       |  614        |  2849      |
|  10x499x499x3       | 250x250       |  2267        |  4304      |
|  10x499x499x1       | 998x998       |   58       |  368        |
|  10x499x499x3       | 998x998       |   170       |  538        |

where M items/s is computed as  `time / (iterations * N * IH * IW * C)`, the higher the better.

### Bilinear

| input_shape (N x IH x IW x C) | output_shape (OH x OW) | Old (M items/s) | New (M items/s) |
| ----- | ---- | ---- | ----- |
|  10x499x499x1       | 250x250       |  1282        |  2362 |
|  10x499x499x3       | 250x250       |  2106        |  3275 |
|  10x499x499x1       | 998x998       |  103        |  182   |
|  10x499x499x3       | 998x998       |  179        |  362   |

where M items/s is computed as  `time / (iterations * N * IH * IW * C)`, the higher the better.